refund, approve refund

+++ main/views.py
@@ -1,36 +1,90 @@
+from datetime import timezone
+from django.contrib import messages
 from django.contrib.auth import login
-from django.contrib.auth.decorators import login_required
+from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.auth.views import LoginView, LogoutView
-from django.utils.decorators import method_decorator
 from django.views import View
 from django.views.generic import ListView
 from django.views.generic.edit import FormView
 from .forms import UserCreationForm, ProductForm
 from django.shortcuts import render, redirect, get_object_or_404
-from .models import Product, Purchase
+from .models import Product, Purchase, Refund
+from django.utils import timezone
 
-@method_decorator(login_required, name='dispatch')
-class BuyProduct(View):
+
+class CreateRefund(LoginRequiredMixin, View):
+    def post(self, request, purchase_id):
+        purchase = get_object_or_404(Purchase, pk=purchase_id)
+
+        if Refund.objects.filter(refund_purchase=purchase).exists():
+            messages.error(request, 'A refund already exists for this purchase.')
+            return redirect('purchase_list')
+
+        time_since_purchase = timezone.localtime(timezone.now()) - purchase.purchase_time
+        if time_since_purchase.total_seconds() > 180:
+            messages.error(request, 'The return period has expired.')
+            return redirect('purchase_list')
+
+        refund = Refund.objects.create(refund_purchase=purchase)
+        return render(request, 'main/refund/confirm_refund.html', {'refund': refund})
+
+
+class RefundList(LoginRequiredMixin, ListView):
+    model = Refund
+    template_name = 'main/refund/refund_list.html'
+
+
+class RefundAgree(LoginRequiredMixin, View):
+    def post(self, request, refund_id):
+        if not request.user.is_staff:
+            messages.error(request, "The administrator must confirm your refund.")
+            return redirect('login')
+
+        refund = get_object_or_404(Refund, pk=refund_id)
+        action = request.POST.get('action')
+
+        if action == 'agree':
+            purchase = refund.refund_purchase
+
+            product = purchase.product
+            product.stock += purchase.quantity
+            product.save()
+
+            purchase.delete()
+            refund.delete()
+        elif action == 'reject':
+            refund.delete()
+
+        return redirect('refunds')
+
+
+class BuyProduct(LoginRequiredMixin, View):
     def post(self, request, product_id):
         product = get_object_or_404(Product, id=product_id)
         quantity = int(request.POST.get('quantity'))
 
         if product.stock < quantity:
-            return render(request, 'main/product/product_list.html', {'eror_message': 'Not enough items in stock.'})
-        if request.user.wallet < product.price * quantity:
+            return render(request, 'main/product/product_list.html', {'error_message': 'Not enough items in stock.'})
+
+        user = request.user
+        total_price = product.price * quantity
+
+        if total_price > request.user.wallet:
+            messages.error(request, "You don't have enough funds for this purchase")
             return render(request, 'main/product/product_list.html', {'error_message': "You don't have enough money."})
+
+        user.wallet -= total_price
+        user.save()
 
         Purchase.objects.create(user=request.user, product=product, quantity=quantity)
 
         product.stock -= quantity
         product.save()
 
-        request.user.wallet -= product.price * quantity
-        request.user.save()
-
         return redirect('product_list')
 
-class PurchaseList(ListView):
+
+class PurchaseList(LoginRequiredMixin, ListView):
     model = Purchase
     template_name = 'main/purchase/purchase_list.html'
     context_object_name = 'purchases'
@@ -39,21 +93,29 @@ class PurchaseList(ListView):
         return Purchase.objects.filter(user=self.request.user)
 
 
-@method_decorator(login_required, name='dispatch')
-class DeleteProduct(View):
+class DeleteProduct(LoginRequiredMixin, View):
     def get(self, request, product_id):
+        if not request.user.is_staff:
+            return redirect('login')
+
         product = get_object_or_404(Product, pk=product_id)
         product.delete()
         return redirect('product_list')
 
-@method_decorator(login_required, name='dispatch')
-class EditProduct(View):
+
+class EditProduct(LoginRequiredMixin, View):
     def get(self, request, product_id):
+        if not request.user.is_staff:
+            return redirect('login')
+
         product = get_object_or_404(Product, id=product_id)
         form = ProductForm(instance=product)
         return render(request, 'main/product/edit_product.html', {'form': form})
 
     def post(self, request, product_id):
+        if not request.user.is_staff:
+            return redirect('login')
+
         product = get_object_or_404(Product, id=product_id)
         form = ProductForm(request.POST, instance=product)
         if form.is_valid():
@@ -61,13 +123,19 @@ class EditProduct(View):
             return redirect('product_list')
         return render(request, 'main/product/edit_product.html', {'form': form})
 
-@method_decorator(login_required, name='dispatch')
-class AddProduct(View):
+
+class AddProduct(LoginRequiredMixin, View):
     def get(self, request):
+        if not request.user.is_staff:
+            return redirect('login')
+
         form = ProductForm()
         return render(request, 'main/product/add_product.html', {'form': form})
 
     def post(self, request):
+        if not request.user.is_staff:
+            return redirect('login')
+
         form = ProductForm(request.POST)
         if form.is_valid():
             form.save()
